# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Crypto Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get bot configuration
      responses:
        "200":
          description: Bot configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    put:
      operationId: updateBotConfig
      tags: [bot]
      summary: Update bot configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigUpdate"
      responses:
        "200":
          description: Updated bot configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get recent bot activity logs
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Bot activity logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/BotLog"
                required: [logs]
components:
  schemas:
...
    BotConfig:
      type: object
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        tradeAmount:
          type: number
        maxPositions:
          type: integer
        stopLossPercent:
          type: number
        takeProfitPercent:
          type: number
        minConfidence:
          type: number
        enabledIndicators:
          type: array
          items:
            type: string
        autoTrade:
          type: boolean
        useAiTargets:
          type: boolean
        checkIntervalSeconds:
          type: integer
      required: [symbol, timeframe, tradeAmount, maxPositions, stopLossPercent, takeProfitPercent, minConfidence, enabledIndicators, autoTrade, useAiTargets, checkIntervalSeconds]
    BotConfigUpdate:
      type: object
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        tradeAmount:
          type: number
        maxPositions:
          type: integer
        stopLossPercent:
          type: number
        takeProfitPercent:
          type: number
        minConfidence:
          type: number
        enabledIndicators:
          type: array
          items:
            type: string
        autoTrade:
          type: boolean
        useAiTargets:
          type: boolean
        checkIntervalSeconds:
          type: integer
    BotLog:
      type: object
      properties:
        id:
          type: string
        level:
          type: string
          enum: ["info", "warning", "error", "trade"]
        message:
          type: string
        symbol:
          type: string
        action:
          type: string
        timestamp:
          type: number
      required: [id, level, message, timestamp]